requirements.txt
# Core ML/AI packages
openai
numpy>=1.21.0
pandas>=1.3.0
transformers>=4.20.0
torch>=1.12.0
scikit-learn>=1.0.0

# Embedding and Retrieval
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.0
bm25s
rankify[full]>=0.1.3

# Data Loading and Processing
datasets>=2.0.0
evaluate>=0.4.0
ir-datasets>=0.5.0

# RAG Framework
langchain>=0.0.200

# Utilities
python-dotenv>=0.19.0
requests>=2.25.1
tqdm>=4.64.0
wandb>=0.13.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Additional dependencies for specific components
accelerate>=0.20.0
sentencepiece>=0.1.90